// FRONTEND/lib/types/database.ts
export type UserRole = "CLIENT" | "PROGRAMMER" | "ADMIN";

export type ProgrammingLanguage =
  | "CSHARP"
  | "NODE_JS"
  | "JAVA"
  | "GO"
  | "PYTHON"
  | "TYPESCRIPT"
  | "JAVASCRIPT"
  | "PHP"
  | "RUST"
  | "KOTLIN"
  | "SWIFT"
  | "OTHER";

export type PlatformType = "WEB" | "DESKTOP" | "MOBILE";
// ...
export interface UserDTO {
  id: string;
  name: string;
  email: string;
  emailVerified: boolean;
  bio?: string | null;
  image?: string | null;
  role?: UserRole | string | null;
  banned?: boolean | null;
  banReason?: string | null;
  banExpires?: string | Date | null;
  createdAt?: string | Date;
  updatedAt?: string | Date;
}
// ...
export interface UserPreferenceDTO {
  id: string;
  userId: string;
  emailNotifications: boolean;
  messageNotifications: boolean;
  projectNotifications: boolean;
  reviewNotifications: boolean;
  createdAt?: string | Date | null;
  updatedAt?: string | Date | null;
}